Add files via upload
train.json was downloaded from NLVR repo at https://github.com/lil-lab/nlvr

text file is an example output which is a dictionary with keys which the user provided at input prompt and the keys have lists with sentences stored such that those sentences have the corresponding key occurring in them.

NLVR_datagen.py is the main python file. It would scan the entire json to find match for the base templates. "There is exactly" is the base search template here. For the rest of the program, it checks wether a new sentence has a recurring string in it. If it is new, the user is asked for an input or else the sentence is added to the corresponding key in the list.